Refactor
Moved services from shared folder to src folder
