+++ src/context/services/ServicesContext.ts
@@ -1,4 +1,4 @@
-import { PeopleService } from '../../shared/services/PeopleService';
+import { PeopleService } from '../../services/PeopleService';
 
 export interface ServicesContext {
 	peopleService: PeopleService;
